Improve build.sh error handling
Return a meaningful error when building samples without
building the root project before.

Closes gh-92

+++ spring-graal-native-samples/springmvc-tomcat/compile.sh
@@ -25,6 +25,11 @@ cp -R META-INF BOOT-INF/classes
 LIBPATH=`find BOOT-INF/lib | tr '\n' ':'`
 CP=BOOT-INF/classes:$LIBPATH:$FEATURE
 
+if [ ! -f "$FEATURE" ]; then
+    printf "${RED}FAILURE${NC}: $FEATURE does not exist, please build the root project before building this sample.\n"
+    exit 1
+fi
+
 echo "Generating reflection files for $ARTIFACT"
 rm -rf graal/META-INF 2>/dev/null
 mkdir -p graal/META-INF/native-image
